Added 2 new custom Serpents Hand Roles.
