+++ SerpentsHand/Roles/SerpentsHandSpecialist.cs
@@ -1,0 +1,111 @@
+﻿using Exiled.API.Enums;
+using Exiled.API.Features;
+using Exiled.API.Features.Attributes;
+using Exiled.API.Features.Spawn;
+using Exiled.CustomRoles.API.Features;
+using Exiled.Events.EventArgs.Player;
+using Exiled.Events.EventArgs.Server;
+using PlayerRoles;
+using System.Collections.Generic;
+using System.ComponentModel;
+
+using PlayerEvent = Exiled.Events.Handlers.Player;
+
+namespace SerpentsHand.Roles
+{
+    [CustomRole(RoleTypeId.Tutorial)]
+    public class SerpentsHandSpecialist : CustomRole
+    {
+        public override uint Id { get; set; } = 3;
+        public override RoleTypeId Role { get; set; } = RoleTypeId.Tutorial;
+        public override int MaxHealth { get; set; } = 100;
+        public override string Name { get; set; } = "Serpents Hand Specialist";
+        public override string Description { get; set; } = "Help the SCPs by killing all other classes";
+        public override string CustomInfo { get; set; } = "Serpents Hand Specialist";
+
+        public override List<string> Inventory { get; set; } = new()
+       {
+          $"{ItemType.GunE11SR}",
+          $"{ItemType.KeycardChaosInsurgency}",
+          $"{ItemType.GrenadeFlash}",
+          $"{ItemType.Radio}",
+          $"{ItemType.Medkit}",
+          $"{ItemType.ArmorCombat}"
+       };
+
+        public override SpawnProperties SpawnProperties { get; set; } = new()
+        {
+            StaticSpawnPoints = new List<StaticSpawnPoint>
+          {
+             new()
+             {
+                Name = "Spawn Point",
+                Position = new UnityEngine.Vector3(63f, 992f, -50f),
+                Chance = 100
+             }
+          }
+        };
+        public override Dictionary<AmmoType, ushort> Ammo { get; set; } = new()
+       {
+          { AmmoType.Nato556, 120 }
+       };
+
+        protected override void SubscribeEvents()
+        {
+            PlayerEvent.EnteringPocketDimension += OnEnteringPocketDimension;
+            PlayerEvent.Hurting += OnHurting;
+            PlayerEvent.Shooting += OnShooting;
+            PlayerEvent.ActivatingGenerator += OnActivatingGenerator;
+            PlayerEvent.ChangingRole += OnChangingRole;
+
+            base.SubscribeEvents();
+        }
+
+        protected override void UnsubscribeEvents()
+        {
+            PlayerEvent.EnteringPocketDimension -= OnEnteringPocketDimension;
+            PlayerEvent.Hurting -= OnHurting;
+            PlayerEvent.Shooting -= OnShooting;
+            PlayerEvent.ActivatingGenerator -= OnActivatingGenerator;
+            PlayerEvent.ChangingRole -= OnChangingRole;
+
+            base.UnsubscribeEvents();
+        }
+
+        private void OnEnteringPocketDimension(EnteringPocketDimensionEventArgs ev)
+        {
+            if (Check(ev.Player))
+                ev.IsAllowed = false;
+        }
+
+        private void OnHurting(HurtingEventArgs ev)
+        {
+            if (ev.Attacker is null) return;
+            if ((Check(ev.Player) || Check(ev.Attacker)) && (ev.Player.IsScp || ev.Attacker.IsScp))
+                ev.IsAllowed = false;
+            //if ((Check(ev.Player) && ev.Attacker.Role.Team == Team.SCPs) ||
+            //    (ev.Attacker != null && Check(ev.Attacker) && ev.Player.Role.Team == Team.SCPs) ||
+            //    (ev.Attacker != null && Check(ev.Attacker) && Check(ev.Player) && ev.Player != ev.Attacker))
+            //	ev.IsAllowed = false;
+        }
+
+        private void OnShooting(ShootingEventArgs ev)
+        {
+            Player target = Player.Get(ev.TargetNetId);
+            if (target != null && target.Role == RoleTypeId.Scp096 && Check(ev.Player))
+                ev.IsAllowed = false;
+        }
+
+        private void OnActivatingGenerator(ActivatingGeneratorEventArgs ev)
+        {
+            if (Check(ev.Player))
+                ev.IsAllowed = false;
+        }
+
+        private void OnChangingRole(ChangingRoleEventArgs ev)
+        {
+            if (Plugin.Instance.Config.SpawnManager.AutoConvertTutorial && ev.NewRole == Role && !ev.Player.IsOverwatchEnabled && !Check(ev.Player))
+                AddRole(ev.Player);
+        }
+    }
+}
